Updating deploy scripts to make use of flyway.

+++ aws/updateStackDbFromLocal.sh
@@ -5,15 +5,17 @@
 source ./devUtil.sh
 
 #location of the dist folder after created locally with gulp build
-sendFile="/home/t/Desktop/DMC/dmcdb"
+sendFile="/home/dmcmillen/development/projects/dmc/dmcdb"
 #key for dbend machine
-db_ssh_keyC="/home/t/Desktop/keys/elkkey.pem"
+db_ssh_keyC="/home/dmcmillen/development/projects/dmc/aws/elkkey.pem"
 #db machine user do not chnage for aws
 db_userC=ec2-user
 #ip of dbend machine
 db_hostC="54.226.135.178"
 PSQLDBNAME="gforge"
 PSQLUSER="gforge"
+PSQLPASS="gforge"
+deploymentEnv="development"
 
 
 scpSend() {
@@ -33,15 +35,23 @@ updatedb() {
     sudo service postgresql94 stop
     sudo service postgresql94 start
 
+
+if [ "$deploymentEnv" = "production" ]; then
+    echo expression evaluated as true
+ # ./flyway migrate info -configFile=conf/core/flyway.conf
+ else
     echo "Dropping $PSQLDBNAME -- db"
     sudo -u postgres psql -c "DROP DATABASE $PSQLDBNAME"
     echo "Create new DB "
     psql -U postgres -c "CREATE DATABASE $PSQLDBNAME WITH OWNER $PSQLUSER;"
     echo "Inserting sample data"
-    psql -U postgres -d gforge < gforge.psql
-    psql -U postgres -d gforge < insert_sample_data.psql
+     
+ ./flyway clean migrate info -configFile=conf/core/flyway.conf -url=jdbc:postgresql://localhost:5432/$PSQLDBNAME  -user=$PSQLUSER -password=$PSQLPASS
+ # load sample data, including DMDII member organizations
+ ./flyway migrate info -configFile=conf/data/flyway.conf -url=jdbc:postgresql://localhost:5432/$PSQLDBNAME  -user=$PSQLUSER -password=$PSQLPASS
     rm -rf /tmp/dmcdb
-    exit
+fi
+exit
 +
 }
